+++ graphdatascience/query_runner/query_runner.py
@@ -9,16 +9,18 @@ from .graph_constructor import GraphConstructor
 
 class QueryRunner(ABC):
     @abstractmethod
-    def run_query(self, query: str, params: Optional[Dict[str, Any]] = None, db: Optional[str] = None) -> DataFrame:
+    def run_query(
+        self, query: str, params: Optional[Dict[str, Any]] = None, database: Optional[str] = None
+    ) -> DataFrame:
         pass
 
     def run_query_with_logging(
-        self, query: str, params: Optional[Dict[str, Any]] = None, db: Optional[str] = None
+        self, query: str, params: Optional[Dict[str, Any]] = None, database: Optional[str] = None
     ) -> DataFrame:
-        return self.run_query(query, params, db)
+        return self.run_query(query, params, database)
 
     @abstractmethod
-    def set_database(self, db: str) -> None:
+    def set_database(self, database: str) -> None:
         pass
 
     def close(self) -> None: